refactor all the things
filtering now happens via the Pipeline class, which makes it easier to
make certain filters conditional.  Moved code block handling into a
filter, which simplifies the main loop.

Rules are now registered in a list, which preserves ordering which is
important when we are running collect_{refs,links} and
missing_ref_link during the atend phase.

Implement include_rules/exclude_rules to control which rules are
applied.

Tests for all new features.